backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.51.2
chromadb>=0.5.0
sentence-transformers==2.3.1
PyMuPDF==1.23.21
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.27.0
supabase==2.4.3
